requirements.txt
filetype==1.2.0
gradio==4.44.0
gradio-rangeslider==0.0.6
numpy==2.1.0
onnx==1.16.1
onnxruntime==1.19.2
opencv-python==4.10.0.84
psutil==6.0.0
tqdm==4.66.5
scipy==1.14.1
grpcio==1.67.0
grpcio-reflection==1.67.0
protobuf==5.27.2
minio==7.2.10
simple-file-checksum==1.2.2
